Simulation with structure

Simulation with structure

Usage

simStructure(
  popStarts = c(2020, 2021, 2022),
  globalNeg = 1,
  NeFunPop,
  samplingDates
)

Arguments

popStarts

Vector of dates when populations start

globalNeg

Value of Ne*g for the global population

NeFunPop

Function determining the demographic trajectory of populations, default is CaveDive model with N=2 and h=0.2

samplingDates

A list containing for each population a vector of sampling dates

Value

A simulated dated phylogeny